+++ backend/canisters/cycles_dispenser/impl/src/lifecycle/post_upgrade.rs
@@ -4,8 +4,10 @@ use crate::Data;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use cycles_dispenser_canister::post_upgrade::Args;
+use ic_cdk::management_canister::{CanisterSettings, LogVisibility, UpdateSettingsArgs};
 use ic_cdk::post_upgrade;
 use stable_memory::get_reader;
+use std::time::Duration;
 use tracing::info;
 
 #[post_upgrade]
@@ -24,4 +26,18 @@ fn post_upgrade(args: Args) {
 
     let total_instructions = ic_cdk::api::call_context_instruction_counter();
     info!(version = %args.wasm_version, total_instructions, "Post-upgrade complete");
+
+    ic_cdk_timers::set_timer(Duration::ZERO, || {
+        ic_cdk::futures::spawn(async {
+            ic_cdk::management_canister::update_settings(&UpdateSettingsArgs {
+                canister_id: ic_cdk::api::canister_self(),
+                settings: CanisterSettings {
+                    log_visibility: Some(LogVisibility::Public),
+                    ..Default::default()
+                },
+            })
+            .await
+            .unwrap()
+        })
+    });
 }