+++ backend/canisters/event_relay/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Log total instructions consumed at end of upgrade ([#7551](https://github.com/open-chat-labs/open-chat/pull/7551))
+- Make canister logs public ([#7675](https://github.com/open-chat-labs/open-chat/pull/7675))
 
 ## [[2.0.1525](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1525-event_relay)] - 2024-12-19